Reshuffle of formatters to be separate from tags so as to allow formatting to occur after inline tag content is concatenated.
Formatters are now passed to the doclet at doclet run time rather than into the tag class. This allows formatting to be applied to raw text after it is concatenated for output.
A markdown formatter has also been added, it requires that PHP Markdown is installed on your system to work.